+++ ui/src/app/utils/aou_external_links.tsx
